If there is only one source configured, shortcut to it
This implements a small optimization shortcut: if there is a single source
defined for completions, lift#complete() will call directly into it, skipping
aggregation of results. Anyway there would be only results from the one source,
so there is not much Lift can do apart from presenting them.

Note that this loses the additional formatting and limiting of maximum results
that Lift does of the called completion sources, so using this shortcut is an
opt-in.